v1.0.2, 서버 응답시 code (에러코드) 함께 반환하도록 추가, 액세스 토큰 만료 시 리프레시 토큰 검사해서 유효하면 새로 발급하도록 추가, 사용자 세션 중복 검사 로직 제거, 채팅방 목록 반환 버그 수정

+++ pkg/utils/board_util.go
@@ -81,7 +81,7 @@ func CheckWriteParameters(c fiber.Ctx) (models.EditorWriteParameter, error) {
 
 	result = models.EditorWriteParameter{
 		BoardUid:    uint(boardUid),
-		UserUid:     actionUserUid,
+		UserUid:     uint(actionUserUid),
 		CategoryUid: uint(categoryUid),
 		Title:       title,
 		Content:     content,
